fix: parameter typo, move module
